datapath/fake: Move FakeDatapath to fake/types package
See previous commit for rationale.
