+++ test/controlplane/services/helpers/lbmap_validation.go
@@ -17,13 +17,13 @@ import (
 	"golang.org/x/exp/constraints"
 	"golang.org/x/exp/maps"
 
-	fakeDatapath "github.com/cilium/cilium/pkg/datapath/fake"
+	fakeTypes "github.com/cilium/cilium/pkg/datapath/fake/types"
 	lb "github.com/cilium/cilium/pkg/loadbalancer"
 	"github.com/cilium/cilium/pkg/testutils/mockmaps"
 	"github.com/cilium/cilium/test/controlplane/suite"
 )
 
-func ValidateLBMapGoldenFile(file string, datapath *fakeDatapath.FakeDatapath) error {
+func ValidateLBMapGoldenFile(file string, datapath *fakeTypes.FakeDatapath) error {
 	lbmap := datapath.LBMockMap()
 	writeLBMap := func() error {
 		f, err := os.OpenFile(file, os.O_CREATE|os.O_TRUNC|os.O_WRONLY, 0644)
